Revert "Switch buffer if file is already opened instead of always call edit."
This reverts commit 8a8fa0948c2732389358985fbbfc0d6b98de2ee9.

The problem that it is supposed to address ("the cursor always returns
at the beginning of the file") is not specific to fswitch and can be
fixed by an autocommand in a more generic way (autocmd BufWinEnter * exe
"normal! g`\"").

On the other hand, this commit introduces two problems:
  - when switching to an unlisted buffer, it remains unlisted. This
    leads, for example, to counterintuitive behavior in buftabline (and
    probably other similar plugins): currently opened buffer is not
    in buffer list.
  - problem with opening symlink buffers (derekwyatt/vim-fswitch#11)

Reverting it solves both.